Move TestGenericDistGit tests to dedicated file
